Commit of the default Drupal Install along with a copy of the database.
Commit of the Drupal installation running with used modules.

I had help from Giselle Zeno, @zeno129 to build what's here already.

Caveat:
Right now the Users on the Database don't have passwords assigned. To generate the password, you need to go to the 'scripts' directory and look for the script to generate passwords. You need to run it, generate a hash and assign it to the users table.

To-Do:
 - I need to generate and export things as features to stream line the process in the future.
 - Theming still needs to be worked.
 - Some rules still need to be created.

Signed-off-by: Ruben Varela <ruben@rubenvarela.com>